+++ packages/node-core/src/indexer/fetch.service.ts
@@ -2,22 +2,21 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import assert from 'assert';
-import util from 'util';
 import {OnApplicationShutdown} from '@nestjs/common';
 import {EventEmitter2} from '@nestjs/event-emitter';
 import {SchedulerRegistry} from '@nestjs/schedule';
-import {BaseDataSource, IProjectNetworkConfig} from '@subql/types-core';
-import {range, without} from 'lodash';
+import {BaseDataSource} from '@subql/types-core';
+import {range} from 'lodash';
 import {NodeConfig} from '../configure';
 import {IndexerEvent} from '../events';
 import {getLogger} from '../logger';
-import {cleanedBatchBlocks, delay, transformBypassBlocks, waitForBatchSize} from '../utils';
+import {delay, filterBypassBlocks, waitForBatchSize} from '../utils';
 import {IBlockDispatcher} from './blockDispatcher';
 import {mergeNumAndBlocksToNums} from './dictionary';
 import {DictionaryService} from './dictionary/dictionary.service';
-import {getBlockHeight, mergeNumAndBlocks} from './dictionary/utils';
+import {mergeNumAndBlocks} from './dictionary/utils';
 import {StoreCacheService} from './storeCache';
-import {Header, IBlock, IProjectService} from './types';
+import {BypassBlocks, Header, IBlock, IProjectService} from './types';
 import {IUnfinalizedBlocksServiceUtil} from './unfinalizedBlocks.service';
 
 const logger = getLogger('FetchService');
@@ -28,7 +27,6 @@ export abstract class BaseFetchService<DS extends BaseDataSource, B extends IBlo
   private _latestBestHeight?: number;
   private _latestFinalizedHeight?: number;
   private isShutdown = false;
-  private bypassBlocks: number[] = [];
 
   // If the chain doesn't have a distinction between the 2 it should return the same value for finalized and best
   protected abstract getFinalizedHeader(): Promise<Header>;
@@ -48,7 +46,6 @@ export abstract class BaseFetchService<DS extends BaseDataSource, B extends IBlo
   constructor(
     private nodeConfig: NodeConfig,
     protected projectService: IProjectService<DS>,
-    protected networkConfig: IProjectNetworkConfig,
     protected blockDispatcher: B,
     protected dictionaryService: DictionaryService<DS, FB>,
     private eventEmitter: EventEmitter2,
@@ -77,31 +74,7 @@ export abstract class BaseFetchService<DS extends BaseDataSource, B extends IBlo
     this.isShutdown = true;
   }
 
-  private updateBypassBlocksFromDatasources(): void {
-    const datasources = this.projectService.getDataSourcesMap().getAll();
-
-    const heights = Array.from(datasources.keys());
-
-    for (let i = 0; i < heights.length - 1; i++) {
-      const currentHeight = heights[i];
-      const nextHeight = heights[i + 1];
-
-      const currentDS = datasources.get(currentHeight);
-      // If the value for the current height is an empty array, then it's a gap
-      if (currentDS && currentDS.length === 0) {
-        this.bypassBlocks.push(...range(currentHeight, nextHeight));
-      }
-    }
-  }
-
   async init(startHeight: number): Promise<void> {
-    this.bypassBlocks = [];
-
-    if (this.networkConfig?.bypassBlocks !== undefined) {
-      this.bypassBlocks = transformBypassBlocks(this.networkConfig.bypassBlocks).filter((blk) => blk >= startHeight);
-    }
-
-    this.updateBypassBlocksFromDatasources();
     const interval = await this.getChainInterval();
 
     await Promise.all([this.getFinalizedBlockHead(), this.getBestBlockHead()]);
@@ -344,49 +317,52 @@ export abstract class BaseFetchService<DS extends BaseDataSource, B extends IBlo
   }
 
   private async enqueueBlocks(enqueuingBlocks: (IBlock<FB> | number)[], latestHeight: number): Promise<void> {
-    const cleanedBatchBlocks = this.filteredBlockBatch(enqueuingBlocks);
+    const cleanedBatchBlocks = filterBypassBlocks<FB>(enqueuingBlocks, [
+      ...this.projectService.bypassBlocks,
+      ...this.getDatasourceBypassBlocks(),
+    ]);
     await this.blockDispatcher.enqueueBlocks(
       cleanedBatchBlocks,
-      this.getLatestBufferHeight(cleanedBatchBlocks, enqueuingBlocks, latestHeight)
+      this.getLatestBufferHeight(enqueuingBlocks, latestHeight)
     );
   }
 
   /**
    *
-   * @param cleanedBatchBlocks
    * @param rawBatchBlocks
    * @param latestHeight
    * @private
    */
-  private getLatestBufferHeight(
-    cleanedBatchBlocks: (IBlock<FB> | number)[],
-    rawBatchBlocks: (IBlock<FB> | number)[],
-    latestHeight: number
-  ): number {
+  private getLatestBufferHeight(rawBatchBlocks: (IBlock<FB> | number)[], latestHeight: number): number {
     // When both BatchBlocks are empty, mean no blocks to enqueue and full synced,
     // we are safe to update latestBufferHeight to this number
-    if (cleanedBatchBlocks.length === 0 && rawBatchBlocks.length === 0) {
+    if (rawBatchBlocks.length === 0) {
       return latestHeight;
     }
-    return Math.max(...mergeNumAndBlocksToNums(cleanedBatchBlocks, rawBatchBlocks));
+    return Math.max(...mergeNumAndBlocksToNums([], rawBatchBlocks));
   }
 
-  private filteredBlockBatch(currentBatchBlocks: (number | IBlock<FB>)[]): (number | IBlock<FB>)[] {
-    if (!this.bypassBlocks.length || !currentBatchBlocks) {
-      return currentBatchBlocks;
-    }
+  /**
+   * If a projects datasources are not continuious we can add add them to the bypass blocks
+   * */
+  private getDatasourceBypassBlocks(): BypassBlocks {
+    const datasources = this.projectService.getDataSourcesMap().getAll();
 
-    const cleanedBatch = cleanedBatchBlocks(this.bypassBlocks, currentBatchBlocks);
+    const heights = Array.from(datasources.keys());
 
-    const pollutedBlocks = this.bypassBlocks.filter(
-      (b) => b < Math.max(...currentBatchBlocks.map((b) => getBlockHeight(b)))
-    );
-    if (pollutedBlocks.length) {
-      // inspect limits the number of logged blocks to 100
-      logger.info(`Bypassing blocks: ${util.inspect(pollutedBlocks, {maxArrayLength: 100})}`);
+    const bypassBlocks: BypassBlocks = [];
+
+    for (let i = 0; i < heights.length - 1; i++) {
+      const currentHeight = heights[i];
+      const nextHeight = heights[i + 1];
+
+      const currentDS = datasources.get(currentHeight);
+      // If the value for the current height is an empty array, then it's a gap
+      if (currentDS?.length === 0) {
+        bypassBlocks.push(`${currentHeight}-${nextHeight - 1}`);
+      }
     }
-    this.bypassBlocks = without(this.bypassBlocks, ...pollutedBlocks);
-    return cleanedBatch;
+    return bypassBlocks;
   }
 
   private nextEndBlockHeight(startBlockHeight: number, scaledBatchSize: number): number {